New class for NormedPublication methods, interface
